+++ src/main/services/app-state.ts
@@ -17,11 +17,12 @@ export interface AppSettings {
     | 'classic'
     | 'dark'
     | 'claude-light'
+    | 'elegant'
     | 'claude-dark'
     | 'monokai-pro-dark'
     | 'monokai-pro-light'
     | 'nord-dark';
-  lightTheme: 'classic' | 'claude-light' | 'monokai-pro-light';
+  lightTheme: 'classic' | 'claude-light' | 'elegant' | 'monokai-pro-light';
   darkTheme: 'dark' | 'claude-dark' | 'monokai-pro-dark' | 'nord-dark';
   contentTheme: string;
   codeTheme: string;